requirements.txt
# SecPipeAI — pinned dependencies for reproducibility
# Python 3.10+ required
# XGBoost CPU-only mode enforced via config (device="cpu")

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
xgboost==2.1.1
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
pyyaml==6.0.2
tqdm==4.66.5
scipy==1.13.1
jinja2==3.1.6
